feat: dialogs

+++ src/app/edit-word/edit-word.component.ts
@@ -2,6 +2,7 @@ import { Component, Inject, ViewChild } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { DictionaryEntry, TauriCommunicationService } from '../service/tauri-communication.service';
 import { MatInput } from '@angular/material/input';
+import { ask } from '@tauri-apps/api/dialog';
 
 export interface EditWordData {
   entry : DictionaryEntry
@@ -24,8 +25,16 @@ export class EditWordComponent {
   isBeingEdited : boolean = false;
 
   removeWordFromDictionary() {
-    this.service.removeWord(this.data.entry.word);
-    this.dialogRef.close();
+    ask("Are you sure you want to remove this word? This operation cannot be undone.", {
+      title: 'Remove Word',
+      type: 'warning'
+    }).then((ok) => {
+      if (!ok) {
+        return;
+      }
+      this.service.removeWord(this.data.entry.word);
+      this.dialogRef.close();
+    })
   }
 
   startEditting() {
